event_machine

+++ lib/sidekick/helpers.rb
@@ -7,53 +7,33 @@ require 'tilt'
 
 module Sidekick::Helpers
 
-  module Util
-    # :linux, :darwin, :other
-    def platform
-      [:linux, :darwin].each do |plf|
-        return plf if Config::CONFIG['target_os'] =~ /#{plf}/i
-      end; :other
-    end
+  # system
 
-    def running?(plf)
-      plf == platform
-    end
-
-    def gem_load?(gemname)
-      @installed ||= begin
-        require gemname
-        true
-      rescue LoadError
-        puts "Please install the #{gemname} gem to enable all functions."
-        false
-      end
-    end
-  end
-
+  require 'sidekick/helpers/util'
   include Util
-  extend Util
-
-  module Notification
-    extend ::Sidekick::Helpers::Util
-    def self.show(message, title='Sidekick')
-      if running?(:linux) and gem_load?('libnotify')
-        Libnotify.show :body => message, :summary => title
-      elsif running?(:darwin) and gem_load?('growl')
-        Growl.notify message, :title => title, :name => 'Sidekick'
-      else
-        puts "Notification: #{title} #{message}"
-      end
-    end
-  end
-
-
 
   def log(str)
     puts ' -> ' + str
   end
 
-  def notify(*args)
-    Notification.show(*args)
+  def stop(*prms)
+    Sidekick.stop(*prms)
+  end
+
+  # notifications
+
+  def notify(message, title='Sidekick')
+    
+    gems = {:linux => 'libnotify', :darwin => 'growl'}
+
+    stop('Notifications not supported.') unless platform_load?(
+                                          gems, 'notifications')
+    case platform
+      when :linux
+        Libnotify.show :body => message, :summary => title
+      when :darwin
+        Growl.notify message, :title => title, :name => 'Sidekick'
+    end
   end
 
   def sh(cmd)
@@ -62,15 +42,10 @@ module Sidekick::Helpers
     result
   end
 
-
-
   def restart_passenger
     FileUtils.touch './tmp/restart.txt'
     log 'restarted passenger'
   end
-
-
-
 
   # watches for changes matching the source glob,
   # compiles using the tilt gem, and saves to
@@ -78,20 +53,19 @@ module Sidekick::Helpers
   def auto_compile(source, target)
     watch(source) do |files|
       files.each do |file|
-        begin
-          t = target.gsub(':name', File.basename(file, '.*'))
-          File.open(t, 'w') do |f|
-            f.write(Tilt.new(file).render)
+        if File.exists?(file)
+          begin
+            t = target.gsub(':name', File.basename(file, '.*'))
+            File.open(t, 'w') do |f|
+              f.write(Tilt.new(file).render)
+            end
+            log "render #{file} => #{t}"
+          rescue Exception => e
+            notify "Error in #{file}:\n#{e}"
           end
-          log "render #{file} => #{t}"
-        rescue Exception => e
-          notify "Error in #{file}:\n#{e}"
         end
       end
     end
   end
-
-
-
 
 end
